Extract ConnectionSession from BackendConnection
